# perf: faster key comparison methods