perf: 2.75x faster ddt_entry_compare()

First 256bits of ddt_key_t is a block checksum, which are expected to be close to
random data. Hence, on average, comparison only needs to look at first few bytes of
the keys. To reduce number of conditional jump instructions, the result is computed
as: sign(memcmp(k1, k2)). Sign of an integer 'a' can be obtained as:
`(0 < a) - (a < 0)` := {-1, 0, 1} , which is computed efficiently.

Synthetic performance evaluation of original and new algorithm over 1G random keys
on 2.6GHz Intel(R) Xeon(R) CPU E5-2660 v3:
```
old		6.85789 s
new		2.49089 s
```

EDIT:
For integer types, comparison can be done as `(k1 > k2) - (k1 < k2)`. Performance gain varies  between 1.1x and 2.5x, depending on compiler version and optimization level.