@ironMann 2016-08-27
Obviously there are more places that would benefit from this. Ideally, `sign()` computation would be moved into avl methods. Then, avl comparator semantics would be relaxed to that of `memcmp()`, if that interface change is acceptable?