@richardelling 2016-08-28
beautiful!

yes, I think there are more places where we can do this sort of optimization. Looking the traces I have handy, vdev_queue_offset_compare() and zfs_range_compare() should be reviewed.