@ironMann 2016-08-28
Added few more places that would benefit from this.