@DeHackEd 2016-08-28
The kernel macro `likely(x)` is meant for the situation where it would be extremely abnormal (less than 1 in a million odds) that `x` were false. A miss here often means a complete L2/3 cache miss, in-place code rewriting, and any other performance penalties the compiler and kernel devs can think of. This is not the place to be using it.