@ironMann 2016-08-28
@DeHackEd Hopefully, the case when keys are different is a most likely to happen. `likely` here is a compiler hint to favour that branch. I've compared different compilers and optimization levels and this gives most consistent results.

> A miss here often means a complete L2/3 cache miss, in-place code rewriting, and any other performance penalties the compiler and kernel devs can think of. 

Can you elaborate on this? There are no cache misses, if keys are close together, due to spatial prefetchers, and certainly no code-rewriting (probes should not be affected by this). It's only a hint for compiler to produce code that agrees with out-of-order and speculative execution of the CPU.