@DeHackEd 2016-08-28
https://github.com/torvalds/linux/blob/master/arch/Kconfig#L52 Maybe I'm confusing this feature for something else. I thought this feature was related to the use of `[un]likely()`.

Behaviour of likely() is <a href="http://stackoverflow.com/questions/109710/likely-unlikely-macros-in-the-linux-kernel-how-do-they-work-whats-their?noredirect=1">documented in other places</a> and the trend is that an instruction cache miss is pretty much guaranteed. Linux kernel convention seems to be to use `likely()` mainly to deal with truly abnormal situations like API usage errors, hardware malfunction, and so on.

But at the end of the day it's hard to argue with the benchmark results.