@ironMann 2016-08-28
`config JUMP_LABEL` seem to be an automorphic code thing, and linux kernel specific. 
It's well known, and I agree,  that programmers are terrible at predicting execution (and testing) of their own code. Instruction cache miss is very likely on improperly placed `likely()` in large functions. 
However, these comparison functions are compiled to 15-20 instructions in total, so that's not gonna happen. Speed increase when using `likely()` hints comes from placing first `retq` instruction immediately after first test (not branching after first != case). Example `space_reftree_compare()`:

Original methods has 2 conditional jump instructions before first return. This makes it hard on CPU branch prediction.
```
000000000008dd70 <space_reftree_compare>:
   8dd75:       48 8b 46 18             mov    0x18(%rsi),%rax
   8dd79:       48 39 47 18             cmp    %rax,0x18(%rdi)
   8dd7d:       55                      push   %rbp
   8dd7e:       48 89 e5                mov    %rsp,%rbp
   8dd81:       72 19                   jb     8dd9c <space_reftree_compare+0x2c>
   8dd83:       b8 01 00 00 00          mov    $0x1,%eax
   8dd88:       77 10                   ja     8dd9a <space_reftree_compare+0x2a>
   8dd8a:       31 c0                   xor    %eax,%eax
   8dd8c:       48 39 f7                cmp    %rsi,%rdi
   8dd8f:       ba ff ff ff ff          mov    $0xffffffff,%edx
   8dd94:       0f 97 c0                seta   %al
   8dd97:       0f 42 c2                cmovb  %edx,%eax
   8dd9a:       5d                      pop    %rbp
   8dd9b:       c3                      retq
   8dd9c:       b8 ff ff ff ff          mov    $0xffffffff,%eax
   8dda1:       5d                      pop    %rbp
   8dda2:       c3                      retq
```

New methods has 1 conditional jump instructions immediately before first return [1]. And since the code after the branch is most likely speculatively executed by the CPU, there won't be any pipeline fill latency (if keys are different).
```
000000000008dd70 <space_reftree_compare>:
   8dd75:       48 8b 56 18             mov    0x18(%rsi),%rdx
   8dd79:       48 8b 4f 18             mov    0x18(%rdi),%rcx
   8dd7d:       31 c0                   xor    %eax,%eax
   8dd7f:       55                      push   %rbp
   8dd80:       48 39 d1                cmp    %rdx,%rcx
   8dd83:       48 89 e5                mov    %rsp,%rbp
   8dd86:       0f 92 c2                setb   %dl
   8dd89:       0f 97 c0                seta   %al
   8dd8c:       0f b6 d2                movzbl %dl,%edx
   8dd8f:       29 d0                   sub    %edx,%eax
   8dd91:       74 02                   je     8dd95 <space_reftree_compare+0x25>     <=== [1]
   8dd93:       5d                      pop    %rbp 
   8dd94:       c3                      retq                                          <=== [1]
   8dd95:       31 c0                   xor    %eax,%eax
   8dd97:       48 39 f7                cmp    %rsi,%rdi
   8dd9a:       0f 92 c2                setb   %dl
   8dd9d:       0f 97 c0                seta   %al
   8dda0:       0f b6 d2                movzbl %dl,%edx
   8dda3:       29 d0                   sub    %edx,%eax
   8dda5:       5d                      pop    %rbp
   8dda6:       c3                      retq
```