@richardelling 2016-08-28
can more of these benefit from inlining?