@ironMann 2016-08-28
@richardelling All of these are used as functors for avl tree operations, and as such, they have to be real functions.