@richardelling 2016-08-28
ok, we can tackle more optimizations later