@behlendorf 2016-08-29
Nice work!  These all LGTM and the testing failures are unrelated.  @richardelling @kernelOfTruth @DeHackEd let me know if I should add you as a reviewer when merging this.