@DeHackEd 2016-08-29
No, leave me out. I just remembered a case of "marked improbable branches suffer harshly when they're wrong" and wanted to make sure all was good.