@richardelling 2016-08-30
LGTM,
I’ve been testing for while and do see reductions in CPU time for the various _compares()