@ironMann 2016-08-30
Alright, I'll see if I missed a good candidate, and try to coalesce the code some more.