@ironMann 2016-08-30
I've finished converting all `_compare()` methods to new comparison macros. Conversion is pretty straightforward. Only 2 instances where I had to deviate from the norm are: `ddt_entry_compare()` and `range_tree_seg_compare()`.

@behlendorf If you prefer, I'll squash these into a single commit.