@behlendorf 2016-08-30
@ironMann please, in this case I think it makes sense to squash all of these very similar optimizations.